HT_23/requirements.txt
celery==5.3.6
Django==5.0
django-autoslug==1.9.9
django-celery-results==2.5.1
django-environ==0.11.2
djangorestframework==3.14.0
drf-spectacular==0.27.1
redis==5.0.1
requests==2.31.0


# celery -A sears_products_scraper worker -l INFO
